Refactor settings surface contracts
