+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -9,7 +9,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `e2e:playwright` | 48 | 7500 | 100 | A+ | `revert.spec.ts` A+<br>`interaction-audit-targets.ts` A+<br>`interaction-audit-fixtures.spec.ts` A+ |
 | `scripts` | 23 | 4114 | 99 | A+ | `native-desktop-smoke.sh` A+<br>`smoke.sh` A+<br>`packaged_window.py` A+ |
 | `source:QuillCodeAgent` | 56 | 7168 | 100 | A+ | `TrustedRouterOAuth.swift` A+<br>`FlailDetector.swift` A+<br>`AgentTypes.swift` A+ |
-| `source:QuillCodeApp` | 436 | 49409 | 100 | A+ | `QuillCodeSettingsSurface.swift` A+<br>`WorkspaceSurface.swift` A+<br>`QuillCodeThreadSidebarSurface.swift` A+ |
+| `source:QuillCodeApp` | 438 | 49422 | 100 | A+ | `WorkspaceSurface.swift` A+<br>`QuillCodeThreadSidebarSurface.swift` A+<br>`RuntimeFactory.swift` A+ |
 | `source:QuillCodeCore` | 31 | 2872 | 99 | A+ | `ModelInfo.swift` A+<br>`AppConfig.swift` A+<br>`AutomationModels.swift` A+ |
 | `source:QuillCodePersistence` | 6 | 372 | 100 | A+ | `ConfigStore.swift` A+<br>`SecretStore.swift` A+<br>`ThreadStore.swift` A+ |
 | `source:QuillCodeSafety` | 7 | 1113 | 99 | A+ | `Safety.swift` A+<br>`StaticSafetyReadOnlyShellPolicy.swift` A+<br>`StaticSafetyDownloadPolicy.swift` A+ |
@@ -22,7 +22,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `test:QuillCodeAppTests` | 214 | 33472 | 99 | A+ | `WorkspaceModelPickerSurfaceIntegrationTests.swift` A+<br>`WorkspaceReviewIntegrationTests.swift` A+<br>`WorkspaceBrowserIntegrationTests.swift` A+ |
 | `test:QuillCodeCoreTests` | 3 | 735 | 98 | A+ | `CoreModelTests.swift` A+<br>`VerificationResultParserTests.swift` A+<br>`ModelTokenUsageTests.swift` A+ |
 | `test:QuillCodeDesktopTests` | 5 | 1328 | 100 | A+ | `QuillCodeDesktopControllerSmokeTests.swift` A+<br>`QuillCodeApprovalNotificationTests.swift` A+<br>`QuillCodeDesktopRenderedSmokeTests.swift` A+ |
-| `test:QuillCodeParityTests` | 145 | 13738 | 100 | A+ | `ParitySwiftInteractionTargetAuditSupport.swift` A+<br>`ParityNativeInteractionContractGateTests.swift` A+<br>`ParitySmokeScriptGateTests.swift` A+ |
+| `test:QuillCodeParityTests` | 145 | 13751 | 100 | A+ | `ParitySwiftInteractionTargetAuditSupport.swift` A+<br>`ParityNativeInteractionContractGateTests.swift` A+<br>`ParitySmokeScriptGateTests.swift` A+ |
 | `test:QuillCodePersistenceTests` | 6 | 385 | 100 | A+ | `JSONProjectStoreTests.swift` A+<br>`ConfigStoreTests.swift` A+<br>`JSONAutomationStoreTests.swift` A+ |
 | `test:QuillCodeSafetyTests` | 5 | 1211 | 100 | A+ | `SafetyGitPullRequestPolicyTests.swift` A+<br>`SafetyDownloadPolicyTests.swift` A+<br>`SafetyPolicyTestCase.swift` A+ |
 | `test:QuillCodeToolsTests` | 32 | 4462 | 100 | A+ | `FileEditSessionGuardTests.swift` A+<br>`FileToolExecutorTests.swift` A+<br>`GitLocalToolExecutorTests.swift` A+ |
@@ -34,7 +34,6 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | --- | ---: | ---: | --- | --- | --- |
 | A+ | 96 | 653 | `test:QuillCodeCoreTests` | `Tests/QuillCodeCoreTests/CoreModelTests.swift` | 29 lines >120 chars |
 | A+ | 96 | 371 | `scripts` | `scripts/native-desktop-smoke.sh` | 22 lines >120 chars; 1 lines >160 chars |
-| A+ | 96 | 340 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSettingsSurface.swift` | 1 lines >120 chars; 42 public declarations |
 | A+ | 96 | 339 | `source:QuillCodeAgent` | `Sources/QuillCodeAgent/TrustedRouterOAuth.swift` | 2 lines >120 chars; 29 public declarations |
 | A+ | 96 | 248 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceSurface.swift` | 1 lines >120 chars; 21 public declarations |
 | A+ | 96 | 235 | `source:QuillCodeAgent` | `Sources/QuillCodeAgent/FlailDetector.swift` | 3 lines >120 chars; 15 public declarations |
@@ -72,6 +71,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 97 | 218 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSidebarItemSurface.swift` | 31 public declarations |
 | A+ | 97 | 206 | `scripts` | `scripts/native_click_probe_contracts/packaged_window.py` | 13 lines >120 chars; 1 lines >160 chars |
 | A+ | 97 | 205 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/Safety.swift` | 19 public declarations |
+| A+ | 97 | 196 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSettingsSurface.swift` | 27 public declarations |
 
 ## Every File
 
@@ -263,6 +263,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 306 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeComposerView.swift` | no automated issues |
 | A+ | 100 | 113 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeComputerUseApprovalSettingsCard.swift` | no automated issues |
 | A+ | 100 | 177 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeComputerUseSettingsCard.swift` | no automated issues |
+| A+ | 100 | 122 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeComputerUseSettingsSurface.swift` | no automated issues |
 | A+ | 100 | 91 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeContextBannerView.swift` | no automated issues |
 | A+ | 100 | 71 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeControlHitTargetViewModifiers.swift` | no automated issues |
 | A+ | 100 | 202 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeDesignSystem.swift` | no automated issues |
@@ -312,7 +313,8 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 275 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSearchAndShortcutDialogs.swift` | no automated issues |
 | A+ | 100 | 35 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSecondaryPanesView.swift` | no automated issues |
 | A+ | 100 | 59 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSettingsDraft.swift` | no automated issues |
-| A+ | 96 | 340 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSettingsSurface.swift` | 1 lines >120 chars; 42 public declarations |
+| A+ | 97 | 196 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSettingsSurface.swift` | 27 public declarations |
+| A+ | 100 | 35 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSettingsUpdate.swift` | no automated issues |
 | A+ | 99 | 153 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSettingsView.swift` | 2 lines >120 chars |
 | A+ | 100 | 38 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSidebarActionsView.swift` | no automated issues |
 | A+ | 100 | 56 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSidebarBulkActionsView.swift` | no automated issues |
@@ -1184,7 +1186,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 99 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceRuntimeToolGateTests.swift` | no automated issues |
 | A+ | 100 | 126 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceSecondaryPaneSurfaceGateTests.swift` | no automated issues |
 | A+ | 100 | 110 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceSettingsSheetGateTests.swift` | no automated issues |
-| A+ | 100 | 21 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceSettingsSurfaceGateTests.swift` | no automated issues |
+| A+ | 100 | 34 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceSettingsSurfaceGateTests.swift` | no automated issues |
 | A+ | 100 | 56 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceSidebarGateTests.swift` | no automated issues |
 | A+ | 100 | 87 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceSidebarPlaywrightGateTests.swift` | no automated issues |
 | A+ | 100 | 26 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceSidebarRowActionGateTests.swift` | no automated issues |
