+++ docs/DECISIONS.md
@@ -319,7 +319,11 @@
 - Transcript message controls live in `QuillCodeTranscriptMessageView.swift`. User and assistant message bubbles, retry/use-as-draft actions, feedback actions, and the shared transcript copy button should evolve there; `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed. Tool-card and artifact-preview rendering remains the next transcript family to extract.
 - Tool-card composition lives in `QuillCodeToolCardView.swift`, while its subfamilies have focused owners: `QuillCodeToolCardControls.swift` owns action rows, status badges, and shared execution-context chips/rails; `QuillCodeToolArtifactViews.swift` owns artifact chips plus text/document/image previews; `QuillCodeToolCardDetailsView.swift` owns raw JSON detail blocks. `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed and how copy actions are wired.
 - Settings sheet ownership is split by role. `QuillCodeSettingsView.swift` owns the sheet shell and authentication controls, `QuillCodeComputerUseSettingsCard.swift` owns Computer Use permission onboarding, `QuillCodeRuntimeIssueView.swift` owns the reusable diagnostics callout, and `QuillCodeSettingsDraft.swift` owns draft-to-update projection. `WorkspaceSwiftUIView.swift` should only present the sheet and apply the resulting settings update.
-- Settings surface contracts live in `QuillCodeSettingsSurface.swift`. `WorkspaceSettingsSurface`, `WorkspaceSettingsUpdate`, Computer Use requirement rows, TrustedRouter sign-in copy, Computer Use permission labels, and compatibility decoding stay beside the settings contract; `WorkspaceSurface.swift` should only carry the aggregate `settings` payload and pass current runtime state into it.
+- Settings surface contracts are split by payload role. `QuillCodeSettingsSurface.swift` owns the aggregate
+  `WorkspaceSettingsSurface` contract and backwards-compatible decoding, `QuillCodeComputerUseSettingsSurface.swift`
+  owns Computer Use requirement rows plus permission/approval copy, and `QuillCodeSettingsUpdate.swift` owns the
+  settings mutation payload. `WorkspaceSurface.swift` should only carry the aggregate `settings` payload and pass
+  current runtime state into it.
 - Terminal and browser utility panes live in `QuillCodeTerminalBrowserPaneView.swift`. Terminal history rows, browser navigation, page snapshots, comments, and pane-local draft state should evolve there; `WorkspaceSwiftUIView.swift` should only decide when the panes are visible and route the resulting actions.
 - Terminal surface contracts live in `QuillCodeTerminalSurface.swift`. `TerminalSurface` and `TerminalCommandSurface` own run/clear availability, cwd labels, command lifecycle labels, and execution-context propagation; `WorkspaceSurface.swift` should only carry the aggregate `terminal` payload and assemble it from terminal state.
 - Secondary utility pane chrome lives in `QuillCodeSecondaryPanesView.swift`. Extensions, Memories, and Automations each have focused native view files: `QuillCodeExtensionsPaneView.swift`, `QuillCodeMemoriesPaneView.swift`, and `QuillCodeAutomationsPaneView.swift`. Automation create-menu routing lives in `QuillCodeAutomationCreateMenu.swift`, workflow card/action rendering lives in `QuillCodeAutomationWorkflowCard.swift`, and shared count/status pills and empty states should evolve in the chrome file; `WorkspaceSwiftUIView.swift` should only decide pane placement and route commands.
